+++ packages/shared/common/src/internal/plugins/index.ts
@@ -1,0 +1,3 @@
+export * from './safeGetHooks';
+export * from './safeGetName';
+export * from './safeRegisterPlugins';